Rename `cmd` to `cmdtools` and fixed some actions command
