+++ .github/workflows/build_image.yml
@@ -41,7 +41,7 @@ jobs:
         uses: docker/setup-buildx-action@v1
       - name: Build base
         uses: docker/build-push-action@v2
-        id: base
+        id: base-dev
         with:
           context: .
           file: Dockerfile-base-dev
